Further split synapse mod.rs into focused submodules (#482)
The previous PR (#495) split the monolithic synapse.rs into submodules
but left mod.rs at 2,485 lines. This extracts three new modules:

- target_analysis.rs (1,020 lines): Per-target analysis loop
- structural_patterns.rs (411 lines): Noisy vs trusted, collapse hidden
- post_processing.rs (315 lines): Impact discounting, sorting, metadata

mod.rs reduced from 2,485 to 860 lines (65% reduction). All files now
under the ~1,500-line project target.

No public API changes. All existing tests pass without modification.
